Move Debug utility from deprecated syntax-definition to runtime-parser.
This will allow to remove syntax-deprecation completely. Debug class is the only one used outside that project.
